investigate response-related effects in motor vertices
I wondered how the x-coordinate effects differed from those that can actually
be attributed to button presses. My first approach was to simply run a
singledot analysis for the motor vertex sources. This showed the largest
effects around 800 ms, as I could expect from the sensor-level average response
effects, but strangely the effects were almost exclusively in the right
hemisphere.

This analysis also included regressors for the x-coordinates of dots 5, 6 and 7
which were supposed to exhibit effects in the considered time window starting
at 700 ms after first dot onset. Because these regressors are correlated
(r~0.3) with the response regressor, it could be that they at least partially
explain away response-related effects. For the 5th dot, for example, I see a
weaker representation of the typical 400 ms past dot onset effect at 800 ms,
but only in the left hemisphere. This fits to the interpretation that the
x-coordinate regressors share effects with the response regressor and
(arbitrarily?) divide up partial effects between them.

To overcome this alignment between dot effects and the response effects I
finally implemented an analysis which investigates response-aligned effects. I
assume that the responses are suitably jittered so that dot effects are
averaged out and correspondingly left out dot regressors from the analysis. The
major issue with this analysis is that it is still based on the original epochs
so that not in all trials data from the time of the response is available.
Specifically, for the time of the response the median number of trials
available per subject is only 146 with 5 subjects even having less than 30
trials, because they had generally long RTs. Nevertheless, I observe strong,
bilateral response effects in this analysis. Interestingly, following my
previous correlation (figure) analysis on the level of brain areas, I see the
largest effects 30 ms after the response. What I didn't expect is that there
are only weak, or no effects building up to the response (e.g. at 50 ms before
the response).

The major finding of this analysis is that the response effects, i.e., increase
in source currents for one response with a decrease in source currents for the
other, mirror the 400 ms dot onset aligned x-coordinate effects. This includes
effects in the eye region. Specifically, it appears that eye region effects in
primary motor cortex precede effects in the hand region, although effects in
the hand region of somatosensory cortices and in premotor areas occur together
with the early eye effects.

The interpretation of these findings is that either a) button presses are
accompanied by eye movements to the corresponding side, b) source localisation
is not precise enough to discriminate the eye and hand regions properly, c) the
preparation and execution of hand movements is in this task coupled to
preparation of eye movements even though they may not be executed, or d)
something else (could subcortical effects be somehow shadowed as eye effects in
cortex in our source reconstruction?).

Results are in:
	source_singledot_201711061631.h5 (dot-aligned, 700 to 890 ms)
	source_singledot_201711081136.h5 (response-aligned, -200 to 50 ms)

+++ show_motor_vertices.py
@@ -9,6 +9,7 @@ Created on Tue Oct 24 15:13:33 2017
 from __future__ import print_function
 import source_visualisations as sv
 import source_statistics as ss
+import re
 import os
 import mne
 
@@ -18,25 +19,51 @@ figdir = os.path.expanduser('~/ZIH/texts/BeeMEG/figures')
 
 
 #%% load data
-r_name = 'dot_x'
+r_name = 'response'
 measure = 'tval'
-exclude_late = False
+make_figures = True
 
-if exclude_late:
-    # vertices of pre-motor and motor areas, baseline (-0.3, 0), first 5 dots, 
-    # data points later than 500 ms before response excluded
-    # trialregs_dot=0, source GLM, sum_dot_y, constregs=0 for 1st dot, 
-    # subject-specific normalisation of DM without centering and scaling by std
-    # label_tc normalised across trials, times and subjects
-    basefile = 'source_sequential_201711031951.h5'
-    latestr = '_exclate'
+# can be:
+# '' for all first-dot-onset-aligned data
+# 'exclate' for first-dot-onset-aligned data excluding data close to response
+# 'ralign' for response aligned data
+datatype = 'ralign'
+
+is_singledot_regressor = lambda name: (
+           (name in ['entropy', 'trial_time', 'response'])
+        or (re.match(r'.*_\d+$', name) is not None))
+
+basefile = None
+if is_singledot_regressor(r_name):
+    if datatype == 'ralign':
+        # vertices of pre-motor and motor areas, baseline (-0.3, 0), dots 5-7, 
+        # response-aligned in [-200, 50], source GLM, 
+        # subject-specific normalisation of DM without centering and scaling by std
+        # label_tc normalised across trials, times and subjects
+        basefile = 'source_singledot_201711081136.h5'
+    elif datatype == '':
+        # vertices of pre-motor and motor areas, baseline (-0.3, 0), dots 5-7, 
+        # time window 0.7-0.89, source GLM, 
+        # subject-specific normalisation of DM without centering and scaling by std
+        # label_tc normalised across trials, times and subjects
+        basefile = 'source_singledot_201711061631.h5'    
 else:
-    # vertices of pre-motor and motor areas, baseline (-0.3, 0), first 5 dots, 
-    # trialregs_dot=0, source GLM, sum_dot_y, constregs=0 for 1st dot, 
-    # subject-specific normalisation of DM without centering and scaling by std
-    # label_tc normalised across trials, times and subjects
-    basefile = 'source_sequential_201710231824.h5'
-    latestr = ''
+    if datatype == 'exclate':
+        # vertices of pre-motor and motor areas, baseline (-0.3, 0), first 5 dots, 
+        # data points later than 500 ms before response excluded
+        # trialregs_dot=0, source GLM, sum_dot_y, constregs=0 for 1st dot, 
+        # subject-specific normalisation of DM without centering and scaling by std
+        # label_tc normalised across trials, times and subjects
+        basefile = 'source_sequential_201711031951.h5'
+    elif datatype == '':
+        # vertices of pre-motor and motor areas, baseline (-0.3, 0), first 5 dots, 
+        # trialregs_dot=0, source GLM, sum_dot_y, constregs=0 for 1st dot, 
+        # subject-specific normalisation of DM without centering and scaling by std
+        # label_tc normalised across trials, times and subjects
+        basefile = 'source_sequential_201710231824.h5'
+
+if len(datatype) > 0 and not datatype.startswith('_'):
+    datatype = '_' + datatype
 
 src_df = ss.load_src_df(basefile, r_name, use_basefile=True)
 ss.add_measure(src_df, 'p_fdr')
@@ -47,13 +74,23 @@ def get_colorinfo(measure, src_df, fdr_alpha=0.01):
     # find measure value that is the first one with p-value equal or smaller 
     # than fdr_alpha
     pdiff = src_df.p_fdr - fdr_alpha
-    fmid = src_df[pdiff <= 0].sort_values('p_fdr')[measure].abs().iloc[-1]
-    
-    if measure == 'tval':
-        colorinfo = {'fmin': fmid / 2., 
-                     'fmid': fmid, 
-                     'fmax': src_df[measure].abs().max(),
-                     'center': 0}
+    try:
+        fmid = src_df[pdiff <= 0].sort_values('p_fdr')[measure].abs().iloc[-1]
+    except IndexError:
+        print('No FDR-corrected significant effects!')
+        if measure == 'tval':
+            fmin = src_df[measure].abs().min()
+            fmax = src_df[measure].abs().max()
+            colorinfo = {'fmin': fmin, 
+                         'fmid': fmax, 
+                         'fmax': fmax + fmax - fmin,
+                         'center': 0}
+    else:
+        if measure == 'tval':
+            colorinfo = {'fmin': fmid / 2., 
+                         'fmid': fmid, 
+                         'fmax': src_df[measure].abs().max(),
+                         'center': 0}
         
     return colorinfo
 
@@ -98,40 +135,41 @@ brain = Brain('fsaverage', 'both', 'inflated', cortex='low_contrast',
 
 add_motor_labels(brain, src_df)
 
-sv.show_label_vertices(src_df, brain, measure, initial_time=400, 
+sv.show_label_vertices(src_df, brain, measure, 
                        **get_colorinfo(measure, src_df))
 
 
 #%% save figures
-#times = [400]
-times = src_df.index.levels[1]
-
-outfiles = []
-for time in times:
-    brain.set_time(time)
+if make_figures:
+    #times = [400]
+    times = src_df.index.levels[1]
     
-    files = {}
-    for hemi, view in views.iteritems():
-        brain.show_view(*view)
-        file = os.path.join(figdir, 'motor_vertices_%s_%s_%s_%d%s.png' 
-                            % (r_name, measure, hemi, time, latestr))
-        files[hemi] = file
-        brain.save_image(file, antialiased=True)
+    outfiles = []
+    for time in times:
+        brain.set_time(time)
         
-    # stitch them together
-    outfiles.append(os.path.join(figdir, 'motor_vertices_%s_%s_%d%s.png' % 
-                                 (r_name, measure, time, latestr)))
-    os.system("montage -tile 2x1 -geometry +0+0 %s %s" % (
-        ' '.join([files['lh'], files['rh']]), outfiles[-1]))
-
-if len(times) > 1:
-    import shutil
-    for ind, file in enumerate(outfiles):
-        shutil.copyfile(file, 'tmp_%d.png' % (ind+1))
+        files = {}
+        for hemi, view in views.iteritems():
+            brain.show_view(*view)
+            file = os.path.join(figdir, 'motor_vertices_%s_%s_%s_%d%s.png' 
+                                % (r_name, measure, hemi, time, datatype))
+            files[hemi] = file
+            brain.save_image(file, antialiased=True)
+            
+        # stitch them together
+        outfiles.append(os.path.join(figdir, 'motor_vertices_%s_%s_%d%s.png' % 
+                                     (r_name, measure, time, datatype)))
+        os.system("montage -tile 2x1 -geometry +0+0 %s %s" % (
+            ' '.join([files['lh'], files['rh']]), outfiles[-1]))
     
-    mvfile = os.path.join(figdir, 'motor_vertices_%s_%s%s.mp4' % 
-                          (r_name, measure, latestr))
-    os.system('avconv -f image2 -r 2 -i tmp_%%d.png -vcodec mpeg4 -y %s' 
-              % mvfile)
-    
-    os.system('rm tmp_*.png')
+    if len(times) > 1:
+        import shutil
+        for ind, file in enumerate(outfiles):
+            shutil.copyfile(file, 'tmp_%d.png' % (ind+1))
+        
+        mvfile = os.path.join(figdir, 'motor_vertices_%s_%s%s.mp4' % 
+                              (r_name, measure, datatype))
+        os.system('avconv -f image2 -r 2 -i tmp_%%d.png -vcodec mpeg4 -y %s' 
+                  % mvfile)
+        
+        os.system('rm tmp_*.png')
